Introduces abstract english visitor.
