+++ src/main/java/com/progressiveaccess/cmlspeech/speech/en/EnSimpleSpeechVisitor.java
@@ -28,13 +28,7 @@
 package com.progressiveaccess.cmlspeech.speech.en;
 
 import com.progressiveaccess.cmlspeech.analysis.RichStructureHelper;
-import com.progressiveaccess.cmlspeech.connection.Bridge;
-import com.progressiveaccess.cmlspeech.connection.BridgeAtom;
-import com.progressiveaccess.cmlspeech.connection.ConnectingBond;
-import com.progressiveaccess.cmlspeech.connection.SharedAtom;
-import com.progressiveaccess.cmlspeech.connection.SharedBond;
-import com.progressiveaccess.cmlspeech.connection.SpiroAtom;
-import com.progressiveaccess.cmlspeech.speech.AbstractSpeechVisitor;
+import com.progressiveaccess.cmlspeech.speech.Language;
 import com.progressiveaccess.cmlspeech.structure.RichAliphaticChain;
 import com.progressiveaccess.cmlspeech.structure.RichAtom;
 import com.progressiveaccess.cmlspeech.structure.RichAtomSet;
@@ -44,45 +38,15 @@ import com.progressiveaccess.cmlspeech.structure.RichFunctionalGroup;
 import com.progressiveaccess.cmlspeech.structure.RichFusedRing;
 import com.progressiveaccess.cmlspeech.structure.RichIsolatedRing;
 import com.progressiveaccess.cmlspeech.structure.RichMolecule;
-import com.progressiveaccess.cmlspeech.structure.RichSetType;
 import com.progressiveaccess.cmlspeech.structure.RichSubRing;
 
 import java.util.Iterator;
-import java.util.SortedSet;
-import java.util.TreeSet;
 
 /**
  * Produces the simple speech for structures.
  */
 
-public class EnSimpleSpeechVisitor extends AbstractSpeechVisitor {
-
-  private boolean shortDescription = false;
-
-
-  @Override
-  public void visit(final RichBond bond) {
-    super.visit(bond);
-    this.addSpeech("bond");
-  }
-
-
-  @Override
-  public void visit(final RichAtom atom) {
-    Integer position = this.getContextPositions().getPosition(atom.getId());
-    // TODO (sorge) Maybe take the supersystem of the atom outside the context.
-    if (position == null) {
-      this.describeSuperSystem(atom);
-      return;
-    }
-    super.visit(atom);
-    this.addSpeech(position);
-    if (this.shortDescription) {
-      return;
-    }
-    this.describeHydrogenBonds(atom);
-  }
-
+public class EnSimpleSpeechVisitor extends EnSpeechVisitor {
 
   @Override
   public void visit(final RichIsolatedRing ring) {
@@ -162,70 +126,6 @@ public class EnSimpleSpeechVisitor extends AbstractSpeechVisitor {
   }
 
 
-  @Override
-  public void visit(final SpiroAtom spiroAtom) {
-    this.shortDescription = true;
-    this.addSpeech("spiro atom");
-    RichStructureHelper.getRichAtom(spiroAtom.getConnector()).accept(this);
-    this.addSpeech("to");
-    RichStructureHelper.getRichAtomSet(spiroAtom.getConnected()).accept(this);
-    this.shortDescription = false;
-  }
-
-
-  @Override
-  public void visit(final BridgeAtom bridgeAtom) {
-    this.addSpeech("bridge atom");
-    RichStructureHelper.getRichAtom(bridgeAtom.getConnector()).accept(this);
-  }
-
-
-  @Override
-  public void visit(final ConnectingBond bond) {
-    RichStructureHelper.getRichBond(bond.getConnector()).accept(this);
-    // TODO (sorge) The past tense here is problematic!
-    this.modSpeech("ed");
-    this.addSpeech("to");
-    this.shortDescription = true;
-    String connected = bond.getConnected();
-    if (RichStructureHelper.isAtom(connected)) {
-      RichStructureHelper.getRichAtom(connected).accept(this);
-    } else {
-      RichStructureHelper.getRichAtomSet(connected).accept(this);
-    }
-    this.shortDescription = false;
-  }
-
-
-  @Override
-  public void visit(final SharedAtom sharedAtom) {
-    this.shortDescription = true;
-    this.addSpeech("shared atom");
-    RichStructureHelper.getRichAtom(sharedAtom.getConnector()).accept(this);
-    this.addSpeech("with");
-    RichStructureHelper.getRichAtomSet(sharedAtom.getConnected()).accept(this);
-    this.shortDescription = false;
-  }
-
-
-  @Override
-  public void visit(final SharedBond sharedBond) {
-    this.addSpeech("shared");
-    RichStructureHelper.getRichBond(sharedBond.getConnector()).accept(this);
-  }
-
-
-  @Override
-  public void visit(final Bridge bridge) {
-    this.shortDescription = true;
-    this.addSpeech("fused with");
-    RichStructureHelper.getRichAtomSet(bridge.getConnected()).accept(this);
-    this.addSpeech("via");
-    bridge.getBridges().forEach(c -> c.accept(this));
-    this.shortDescription = false;
-  }
-
-
   // TODO (sorge) For the following utility functions, see if they can be
   // refactored with walk methods, etc.
   private void describeReplacements(final RichAtomSet system) {
@@ -235,7 +135,7 @@ public class EnSimpleSpeechVisitor extends AbstractSpeechVisitor {
       final RichAtom atom = RichStructureHelper.getRichAtom(value);
       if (!atom.isCarbon()) {
         this.addSpeech("with");
-        super.visit(atom);
+        this.addSpeech(Language.getAtomTable().lookup(atom));
         this.addSpeech("at position");
         this.addSpeech(system.getPosition(value));
       }
@@ -266,72 +166,6 @@ public class EnSimpleSpeechVisitor extends AbstractSpeechVisitor {
       this.addSpeech("and");
       this.addSpeech(atomB);
     }
-  }
-
-
-  private void describeSubstitutions(final RichAtomSet system) {
-    final SortedSet<Integer> subst = new TreeSet<Integer>();
-    for (final String atom : system.getConnectingAtoms()) {
-      subst.add(system.getPosition(atom));
-    }
-    switch (subst.size()) {
-      case 0:
-        return;
-      case 1:
-        this.addSpeech("Substitution at position");
-        this.addSpeech(subst.iterator().next());
-        return;
-      default:
-        this.addSpeech("Substitutions at positions");
-        for (final Integer position : subst) {
-          this.addSpeech(position);
-          this.addSpeech("and");
-        }
-        this.remSpeech();
-    }
-  }
-
-
-  /** 
-   * Adds description of hydrogen bonds of an atom.
-   * 
-   * @param atom
-   *          The atom to describe.
-   */
-  private void describeHydrogenBonds(RichAtom atom) {
-    final Integer count = atom.getStructure().getImplicitHydrogenCount();
-    switch (count) {
-      case 0:
-        return;
-      case 1:
-        this.addSpeech("bonded to");
-        this.addSpeech(count.toString());
-        this.addSpeech("hydrogen");
-        return;
-      default:
-        this.addSpeech("bonded to");
-        this.addSpeech(count.toString());
-        this.addSpeech("hydrogens");
-        return;
-    }
-  }
-
-
-  private void describeSuperSystem(final RichAtom atom) {
-    this.shortDescription = true;
-    for (String context : atom.getContexts()) {
-      if (RichStructureHelper.isAtomSet(context)) {
-        RichAtomSet set = RichStructureHelper.getRichAtomSet(context);
-        RichSetType type = set.getType();
-        if (type == RichSetType.FUNCGROUP
-            || type == RichSetType.ISOLATED
-            || type == RichSetType.FUSED
-            || type == RichSetType.ALIPHATIC) {
-          set.accept(this);
-        }
-      }
-    }
-    this.shortDescription = false;
   }
 
 }
