+++ src/main/java/com/progressiveaccess/cmlspeech/speech/ja/JaExpertSpeechVisitor.java
@@ -35,6 +35,7 @@ import com.progressiveaccess.cmlspeech.connection.SharedAtom;
 import com.progressiveaccess.cmlspeech.connection.SharedBond;
 import com.progressiveaccess.cmlspeech.connection.SpiroAtom;
 import com.progressiveaccess.cmlspeech.speech.AbstractSpeechVisitor;
+import com.progressiveaccess.cmlspeech.speech.Language;
 import com.progressiveaccess.cmlspeech.structure.RichAliphaticChain;
 import com.progressiveaccess.cmlspeech.structure.RichAtom;
 import com.progressiveaccess.cmlspeech.structure.RichAtomSet;
@@ -48,8 +49,6 @@ import com.progressiveaccess.cmlspeech.structure.RichSubRing;
 
 import com.google.common.base.Joiner;
 
-import java.util.HashMap;
-import java.util.Map;
 import java.util.SortedSet;
 import java.util.TreeSet;
 
@@ -65,7 +64,7 @@ public class JaExpertSpeechVisitor extends AbstractSpeechVisitor {
   
   @Override
   public void visit(final RichBond bond) {
-    super.visit(bond);
+    this.addSpeech(Language.getBondTable().order(bond));
     this.addSpeech("結合"); // bond
   }
 
@@ -78,7 +77,7 @@ public class JaExpertSpeechVisitor extends AbstractSpeechVisitor {
       this.describeSuperSystem(atom);
       return;
     }
-    super.visit(atom);
+    this.addSpeech(Language.getAtomTable().lookup(atom));
     this.addSpeech(position);
     if (this.subject) {
       this.addSpeech("は、"); // Separator (only after subject).
